Extracting a common metadata mapper for work type
The rules for data export is made complicated by our very fluid models
in Curate. However this is a beginning for breaking that apart.
